+++ EssentialFeed/EssentialFeed/FeedCache/LocalFeedItem.swift
@@ -1,0 +1,22 @@
+//
+//  LocalFeedItem.swift
+//  EssentialFeed
+//
+//  Created by Radosław Serek on 12/11/2022.
+//
+
+import Foundation
+
+public struct LocalFeedItem: Equatable {
+    public let id: UUID
+    public let description: String?
+    public let location: String?
+    public let imageURL: URL
+    
+    public init(id: UUID, description: String?, location: String?, imageURL: URL) {
+        self.id = id
+        self.description = description
+        self.location = location
+        self.imageURL = imageURL
+    }
+}
